refactor: memory usage (#1238 #1276 #1277)
